+++ Cargo.toml
@@ -1,19 +1,18 @@
 [package]
-name = "crypto"
+name = "concordium-wallet-crypto-uniffi"
 version = "1.0.0"
 edition = "2021"
 
 [lib]
 crate-type = ["staticlib", "lib"]
-name = "crypto"
 
 [dependencies]
 thiserror = "1"
-uniffi = { version = "0.25", features = ["cli"] }
+uniffi = { version = "0.26", features = ["cli"] }
 wallet_library = { path = "./concordium-base/rust-src/wallet_library" }
 
 [build-dependencies]
-uniffi = { version = "0.25", features = ["build"] }
+uniffi = { version = "0.26", features = ["build"] }
 
 [[bin]]
 name = "uniffi-bindgen"
